{{cookiecutter.project_directory}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_directory}}"
version = "0.1.0"
description = ""
authors = ["GGGoingdown <eddiepai0401@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
gunicorn = "^20.1.0"
sentry-sdk = "^1.19.1"
dependency-injector = "^4.41.0"
loguru = "^0.7.0"
aiohttp = "^3.8.4"
aiohttp-retry = "^2.8.3"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.0"
httpx = "^0.23.3"
pytest-anyio = "^0.0.0"
pytest-asyncio = "^0.21.0"
trio = "^0.22.0"
pytest-trio = "^0.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
